PIP Survey Metadata

Update or load a dataset with survey metadata.

Usage

aux_metadata(
  action = c("update", "load"),
  force = FALSE,
  owner = getOption("pipfun.ghowner"),
  maindir = getOption("pipaux.working_dir"),
  branch = paste0(wrk_release$release, "_", wrk_release$identity),
  tag = NULL,
  detail = getOption("pipaux.detail.raw")
)

Arguments

action

character: Either "load" or "update". Default is "update". If "update" data will be updated on the system. If "load" data is loaded in memory.

force

logical: If TRUE data will be overwritten.

owner

character: Github repo owner. Default is getOption("pipfun.ghowner")

maindir

character: Main directory of project.

branch

GitHub branch to use. By default, this is the release branch set in your R session. To ensure correct behavior, call pipfun::setup_working_release() once per session, followed by pipfun::get_wrk_release() to retrieve the active release. These steps ensure the function uses the correct release branch (e.g., "release_2024Q1"). Alternatively, you can explicitly specify other branches among "DEV", "PROD", or "main" if needed

tag

character: specific release to be used in the update.

detail

has an option TRUE/FALSE, default value is FALSE